+++ lib/src/platform_app.dart
@@ -6,10 +6,17 @@
 
 import 'package:flutter/cupertino.dart' show CupertinoApp, CupertinoThemeData;
 import 'package:flutter/material.dart'
-    show MaterialApp, ScaffoldMessengerState, Theme, ThemeData, ThemeMode;
+    show
+        MaterialApp,
+        ScaffoldMessengerState,
+        Theme,
+        ThemeData,
+        ThemeMode,
+        kThemeAnimationDuration;
 import 'package:flutter/widgets.dart';
 
 import 'platform.dart';
+import 'platform_theme.dart';
 import 'widget_base.dart';
 
 abstract class _BaseData {
@@ -107,6 +114,7 @@ abstract class _BaseRouterData {
     this.routeInformationProvider,
     this.routeInformationParser,
     this.routerDelegate,
+    this.routerConfig,
     this.backButtonDispatcher,
     // ignore: unused_element
     this.restorationScopeId,
@@ -143,6 +151,9 @@ abstract class _BaseRouterData {
 
   /// {@macro flutter.widgets.widgetsApp.routerDelegate}
   final RouterDelegate<Object>? routerDelegate;
+
+  /// {@macro flutter.widgets.widgetsApp.routerConfig}
+  final RouterConfig<Object>? routerConfig;
 
   /// {@macro flutter.widgets.widgetsApp.backButtonDispatcher}
   final BackButtonDispatcher? backButtonDispatcher;
@@ -189,6 +200,8 @@ class MaterialAppData extends _BaseData {
     this.darkTheme,
     this.themeMode,
     this.scaffoldMessengerKey,
+    this.themeAnimationCurve = Curves.linear,
+    this.themeAnimationDuration = kThemeAnimationDuration,
   });
 
   final ThemeData? theme;
@@ -196,6 +209,8 @@ class MaterialAppData extends _BaseData {
   final ThemeData? darkTheme;
   final ThemeMode? themeMode;
   final GlobalKey<ScaffoldMessengerState>? scaffoldMessengerKey;
+  final Curve themeAnimationCurve;
+  final Duration themeAnimationDuration;
 }
 
 class MaterialAppRouterData extends _BaseRouterData {
@@ -215,6 +230,10 @@ class MaterialAppRouterData extends _BaseRouterData {
     super.checkerboardOffscreenLayers,
     super.showSemanticsDebugger,
     super.debugShowCheckedModeBanner,
+    super.routeInformationProvider,
+    super.routeInformationParser,
+    super.routerDelegate,
+    super.routerConfig,
     super.shortcuts,
     super.actions,
     super.onGenerateInitialRoutes,
@@ -225,6 +244,8 @@ class MaterialAppRouterData extends _BaseRouterData {
     this.darkTheme,
     this.themeMode,
     this.scaffoldMessengerKey,
+    this.themeAnimationCurve = Curves.linear,
+    this.themeAnimationDuration = kThemeAnimationDuration,
   });
 
   final ThemeData? theme;
@@ -232,6 +253,8 @@ class MaterialAppRouterData extends _BaseRouterData {
   final ThemeData? darkTheme;
   final ThemeMode? themeMode;
   final GlobalKey<ScaffoldMessengerState>? scaffoldMessengerKey;
+  final Curve themeAnimationCurve;
+  final Duration themeAnimationDuration;
 }
 
 class CupertinoAppData extends _BaseData {
@@ -292,6 +315,7 @@ class CupertinoAppRouterData extends _BaseRouterData {
     super.routeInformationProvider,
     super.routeInformationParser,
     super.routerDelegate,
+    super.routerConfig,
     super.backButtonDispatcher,
     super.scrollBehavior,
     super.useInheritedMediaQuery,
@@ -342,6 +366,9 @@ class PlatformApp extends PlatformWidgetBase<CupertinoApp, MaterialApp> {
   /// {@macro flutter.widgets.widgetsApp.routerDelegate}
   final RouterDelegate<Object>? routerDelegate;
 
+  /// {@macro flutter.widgets.widgetsApp.routerConfig}
+  final RouterConfig<Object>? routerConfig;
+
   /// {@macro flutter.widgets.widgetsApp.backButtonDispatcher}
   final BackButtonDispatcher? backButtonDispatcher;
 
@@ -386,6 +413,7 @@ class PlatformApp extends PlatformWidgetBase<CupertinoApp, MaterialApp> {
   })  : routeInformationProvider = null,
         routeInformationParser = null,
         routerDelegate = null,
+        routerConfig = null,
         backButtonDispatcher = null,
         materialRouter = null,
         cupertinoRouter = null;
@@ -395,6 +423,7 @@ class PlatformApp extends PlatformWidgetBase<CupertinoApp, MaterialApp> {
     this.routeInformationProvider,
     this.routeInformationParser,
     this.routerDelegate,
+    this.routerConfig,
     this.backButtonDispatcher,
     this.widgetKey,
     this.builder,
@@ -426,8 +455,8 @@ class PlatformApp extends PlatformWidgetBase<CupertinoApp, MaterialApp> {
         onUnknownRoute = null,
         routes = null,
         initialRoute = null,
-        this.material = null,
-        this.cupertino = null,
+        material = null,
+        cupertino = null,
         materialRouter = material,
         cupertinoRouter = cupertino;
 
@@ -436,28 +465,38 @@ class PlatformApp extends PlatformWidgetBase<CupertinoApp, MaterialApp> {
     final dataRouter = materialRouter?.call(context, platform(context));
 
     if (routeInformationParser != null ||
-        dataRouter?.routeInformationParser != null) {
-      assert(dataRouter?.routerDelegate != null || routerDelegate != null);
+        dataRouter?.routeInformationParser != null ||
+        routerConfig != null ||
+        dataRouter?.routerConfig != null) {
+      assert(dataRouter?.routerDelegate != null ||
+          routerDelegate != null ||
+          dataRouter?.routerConfig != null ||
+          routerConfig != null);
 
       return MaterialApp.router(
         routeInformationProvider:
             dataRouter?.routeInformationProvider ?? routeInformationProvider,
         routeInformationParser:
-            dataRouter?.routeInformationParser ?? routeInformationParser!,
-        routerDelegate: dataRouter?.routerDelegate ?? routerDelegate!,
+            dataRouter?.routeInformationParser ?? routeInformationParser,
+        routerDelegate: dataRouter?.routerDelegate ?? routerDelegate,
+        routerConfig: dataRouter?.routerConfig ?? routerConfig,
         backButtonDispatcher:
             dataRouter?.backButtonDispatcher ?? backButtonDispatcher,
         builder: dataRouter?.builder ?? builder,
         title: dataRouter?.title ?? title ?? '',
         onGenerateTitle: dataRouter?.onGenerateTitle ?? onGenerateTitle,
         color: dataRouter?.color ?? color,
-        theme: (dataRouter?.theme ?? Theme.of(context))
+        theme: (dataRouter?.theme ??
+                _getMaterialLightThemeData(context) ??
+                Theme.of(context))
             .copyWith(platform: TargetPlatform.android),
-        darkTheme:
-            dataRouter?.darkTheme?.copyWith(platform: TargetPlatform.android),
+        darkTheme: (dataRouter?.darkTheme ?? _getMaterialDarkThemeData(context))
+            ?.copyWith(platform: TargetPlatform.android),
+        themeMode: dataRouter?.themeMode ??
+            _getMaterialThemeMode(context) ??
+            ThemeMode.system,
         highContrastDarkTheme: dataRouter?.highContrastDarkTheme,
         highContrastTheme: dataRouter?.highContrastTheme,
-        themeMode: dataRouter?.themeMode ?? ThemeMode.system,
         locale: dataRouter?.locale ?? locale,
         localizationsDelegates:
             dataRouter?.localizationsDelegates ?? localizationsDelegates,
@@ -495,6 +534,9 @@ class PlatformApp extends PlatformWidgetBase<CupertinoApp, MaterialApp> {
         useInheritedMediaQuery: dataRouter?.useInheritedMediaQuery ??
             useInheritedMediaQuery ??
             false,
+        themeAnimationCurve: dataRouter?.themeAnimationCurve ?? Curves.linear,
+        themeAnimationDuration:
+            dataRouter?.themeAnimationDuration ?? kThemeAnimationDuration,
       );
     } else {
       final data = material?.call(context, platform(context));
@@ -536,11 +578,16 @@ class PlatformApp extends PlatformWidgetBase<CupertinoApp, MaterialApp> {
         debugShowCheckedModeBanner: data?.debugShowCheckedModeBanner ??
             debugShowCheckedModeBanner ??
             true,
-        theme: (data?.theme ?? Theme.of(context))
-            .copyWith(platform: TargetPlatform.android),
         debugShowMaterialGrid: data?.debugShowMaterialGrid ?? false,
-        darkTheme: data?.darkTheme?.copyWith(platform: TargetPlatform.android),
-        themeMode: data?.themeMode ?? ThemeMode.system,
+        theme: (data?.theme ??
+                _getMaterialLightThemeData(context) ??
+                Theme.of(context))
+            .copyWith(platform: TargetPlatform.android),
+        darkTheme: (data?.darkTheme ?? _getMaterialDarkThemeData(context))
+            ?.copyWith(platform: TargetPlatform.android),
+        themeMode: data?.themeMode ??
+            _getMaterialThemeMode(context) ??
+            ThemeMode.system,
         shortcuts: data?.shortcuts ?? shortcuts,
         actions: data?.actions ?? actions,
         onGenerateInitialRoutes:
@@ -552,6 +599,9 @@ class PlatformApp extends PlatformWidgetBase<CupertinoApp, MaterialApp> {
         scrollBehavior: data?.scrollBehavior ?? scrollBehavior,
         useInheritedMediaQuery:
             data?.useInheritedMediaQuery ?? useInheritedMediaQuery ?? false,
+        themeAnimationCurve: data?.themeAnimationCurve ?? Curves.linear,
+        themeAnimationDuration:
+            data?.themeAnimationDuration ?? kThemeAnimationDuration,
       );
     }
   }
@@ -561,18 +611,24 @@ class PlatformApp extends PlatformWidgetBase<CupertinoApp, MaterialApp> {
     final dataRouter = cupertinoRouter?.call(context, platform(context));
 
     if (routeInformationParser != null ||
-        dataRouter?.routeInformationParser != null) {
-      assert(dataRouter?.routerDelegate != null || routerDelegate != null);
+        dataRouter?.routeInformationParser != null ||
+        routerConfig != null ||
+        dataRouter?.routerConfig != null) {
+      assert(dataRouter?.routerDelegate != null ||
+          routerDelegate != null ||
+          dataRouter?.routerConfig != null ||
+          routerConfig != null);
 
       return CupertinoApp.router(
         routeInformationProvider:
             dataRouter?.routeInformationProvider ?? routeInformationProvider,
         routeInformationParser:
-            dataRouter?.routeInformationParser ?? routeInformationParser!,
-        routerDelegate: dataRouter?.routerDelegate ?? routerDelegate!,
+            dataRouter?.routeInformationParser ?? routeInformationParser,
+        routerDelegate: dataRouter?.routerDelegate ?? routerDelegate,
+        routerConfig: dataRouter?.routerConfig ?? routerConfig,
         backButtonDispatcher:
             dataRouter?.backButtonDispatcher ?? backButtonDispatcher,
-        theme: dataRouter?.theme,
+        theme: dataRouter?.theme ?? _getCupertinoTheme(context),
         builder: dataRouter?.builder ?? builder,
         title: dataRouter?.title ?? title ?? '',
         onGenerateTitle: dataRouter?.onGenerateTitle ?? onGenerateTitle,
@@ -615,6 +671,7 @@ class PlatformApp extends PlatformWidgetBase<CupertinoApp, MaterialApp> {
       );
     } else {
       final data = cupertino?.call(context, platform(context));
+
       return CupertinoApp(
         key: data?.widgetKey ?? widgetKey,
         navigatorKey: data?.navigatorKey ?? navigatorKey,
@@ -653,7 +710,7 @@ class PlatformApp extends PlatformWidgetBase<CupertinoApp, MaterialApp> {
         debugShowCheckedModeBanner: data?.debugShowCheckedModeBanner ??
             debugShowCheckedModeBanner ??
             true,
-        theme: data?.theme,
+        theme: data?.theme ?? _getCupertinoTheme(context),
         shortcuts: data?.shortcuts ?? shortcuts,
         actions: data?.actions ?? actions,
         onGenerateInitialRoutes:
@@ -664,5 +721,31 @@ class PlatformApp extends PlatformWidgetBase<CupertinoApp, MaterialApp> {
             data?.useInheritedMediaQuery ?? useInheritedMediaQuery ?? false,
       );
     }
+  }
+
+  ThemeData? _getMaterialLightThemeData(BuildContext context) {
+    return PlatformTheme.of(context)?.materialLightTheme;
+  }
+
+  ThemeData? _getMaterialDarkThemeData(BuildContext context) {
+    return PlatformTheme.of(context)?.materialDarkTheme;
+  }
+
+  ThemeMode? _getMaterialThemeMode(BuildContext context) {
+    return PlatformTheme.of(context)?.themeMode;
+  }
+
+  CupertinoThemeData? _getCupertinoTheme(BuildContext context) {
+    final isDark = PlatformTheme.of(context)?.isDark;
+    final lightTheme = PlatformTheme.of(context)?.cupertinoLightTheme;
+    final darkTheme = PlatformTheme.of(context)?.cupertinoDarkTheme ??
+        PlatformTheme.of(context)?.cupertinoLightTheme;
+    final theme = isDark == null
+        ? null
+        : isDark
+            ? darkTheme
+            : lightTheme;
+
+    return theme;
   }
 }