+++ lib/src/platform_app_bar.dart
@@ -13,6 +13,7 @@ import 'extensions.dart';
 import 'platform.dart';
 import 'platform_provider.dart';
 import 'widget_base.dart';
+import 'parent_widget_finder.dart';
 
 //the default has alpha which will cause the content to slide under the header for ios
 const Color _kDefaultNavBarBorderColor = const Color(0x4C000000);
@@ -70,6 +71,7 @@ class MaterialAppBarData extends _BaseData {
     this.toolbarTextStyle,
     this.scrolledUnderElevation,
     this.surfaceTintColor,
+    this.notificationPredicate,
   });
 
   final List<Widget>? actions;
@@ -95,6 +97,7 @@ class MaterialAppBarData extends _BaseData {
   final TextStyle? toolbarTextStyle;
   final double? scrolledUnderElevation;
   final Color? surfaceTintColor;
+  final ScrollNotificationPredicate? notificationPredicate;
 }
 
 class CupertinoNavigationBarData extends _BaseData {
@@ -188,6 +191,8 @@ class PlatformAppBar
       toolbarTextStyle: data?.toolbarTextStyle,
       scrolledUnderElevation: data?.scrolledUnderElevation,
       surfaceTintColor: data?.surfaceTintColor,
+      notificationPredicate:
+          data?.notificationPredicate ?? defaultScrollNotificationPredicate,
       // deprecated...
       // brightness
       // textTheme
@@ -216,7 +221,7 @@ class PlatformAppBar
     if (heroTag != null) {
       return CupertinoNavigationBar(
         key: data?.widgetKey ?? widgetKey,
-        middle: data?.title ?? title,
+        middle: _getMiddleCupertinoWidget(context, data),
         backgroundColor: data?.backgroundColor ?? backgroundColor,
         automaticallyImplyLeading: data?.automaticallyImplyLeading ??
             automaticallyImplyLeading ??
@@ -225,8 +230,12 @@ class PlatformAppBar
         previousPageTitle: data?.previousPageTitle,
         padding: data?.padding,
         border: data?.border ?? _kDefaultNavBarBorder,
-        leading: (data?.leading ?? leading)?.withMaterial(useMaterial),
-        trailing: (data?.trailing ?? trailing)?.withMaterial(useMaterial),
+        leading: _getLeadingCupertinoWidget(context, data)
+            ?.withMaterial(useMaterial)
+            .withWidgetFinder<CupertinoNavigationBar>(),
+        trailing: (data?.trailing ?? trailing)
+            ?.withMaterial(useMaterial)
+            .withWidgetFinder<CupertinoNavigationBar>(),
         transitionBetweenRoutes: data?.transitionBetweenRoutes ?? true,
         brightness: data?.brightness,
         heroTag: heroTag,
@@ -235,7 +244,7 @@ class PlatformAppBar
 
     return CupertinoNavigationBar(
       key: data?.widgetKey ?? widgetKey,
-      middle: data?.title ?? title,
+      middle: _getMiddleCupertinoWidget(context, data),
       backgroundColor: data?.backgroundColor ?? backgroundColor,
       automaticallyImplyLeading:
           data?.automaticallyImplyLeading ?? automaticallyImplyLeading ?? true,
@@ -243,11 +252,78 @@ class PlatformAppBar
       previousPageTitle: data?.previousPageTitle,
       padding: data?.padding,
       border: data?.border ?? _kDefaultNavBarBorder,
-      leading: (data?.leading ?? leading)?.withMaterial(useMaterial),
-      trailing: (data?.trailing ?? trailing)?.withMaterial(useMaterial),
+      leading: _getLeadingCupertinoWidget(context, data)
+          ?.withMaterial(useMaterial)
+          .withWidgetFinder<CupertinoNavigationBar>(),
+      trailing: (data?.trailing ?? trailing)
+          ?.withMaterial(useMaterial)
+          .withWidgetFinder<CupertinoNavigationBar>(),
       transitionBetweenRoutes: data?.transitionBetweenRoutes ?? true,
       brightness: data?.brightness,
       //heroTag: , used above
     );
+  }
+
+  Widget? _getLeadingCupertinoWidget(
+    BuildContext context,
+    CupertinoNavigationBarData? data,
+  ) {
+    final leadingLocal = data?.leading ?? leading;
+
+    return leadingLocal;
+
+    // Currently there is this issue which prevents this from being done properly
+    // https://github.com/flutter/flutter/issues/89888
+
+    // final ModalRoute<dynamic>? currentRoute = ModalRoute.of(context);
+    // final canPop = currentRoute?.canPop ?? false;
+
+    // if (!canPop) {
+    //   return leadingLocal;
+    // }
+
+    // final useMediaQueryWrapper = PlatformProvider.of(context)
+    //         ?.settings
+    //         .wrapCupertinoAppBarMiddleWithMediaQuery ??
+    //     true;
+
+    // if (!useMediaQueryWrapper) {
+    //   return middleLocal;
+    // }
+
+    // final leadingWithMediaQuery = MediaQuery(
+    //   data: MediaQueryData(
+    //       textScaleFactor: MediaQuery.textScaleFactorOf(context)),
+    //   child: leadingLocal ?? CupertinoNavigationBarBackButton(),
+    // );
+
+    // return leadingWithMediaQuery;
+  }
+
+  Widget? _getMiddleCupertinoWidget(
+    BuildContext context,
+    CupertinoNavigationBarData? data,
+  ) {
+    final middleLocal = data?.title ?? title;
+    if (middleLocal == null) {
+      return null;
+    }
+
+    final useMediaQueryWrapper = PlatformProvider.of(context)
+            ?.settings
+            .wrapCupertinoAppBarMiddleWithMediaQuery ??
+        true;
+
+    if (!useMediaQueryWrapper) {
+      return middleLocal;
+    }
+
+    final middleWithMediaQuery = MediaQuery(
+      data: MediaQueryData(
+          textScaleFactor: MediaQuery.textScaleFactorOf(context)),
+      child: middleLocal,
+    );
+
+    return middleWithMediaQuery;
   }
 }