+++ lib/src/platform_icon_button.dart
@@ -4,12 +4,16 @@
  * See LICENSE for distribution and usage details.
  */
 
-import 'package:flutter/cupertino.dart' show CupertinoButton, CupertinoColors;
-import 'package:flutter/material.dart' show IconButton, VisualDensity;
+import 'package:flutter/cupertino.dart'
+    show CupertinoButton, CupertinoColors, CupertinoNavigationBar;
+import 'package:flutter/material.dart'
+    show IconButton, VisualDensity, ButtonStyle;
 import 'package:flutter/widgets.dart';
+import 'package:flutter_platform_widgets/src/parent_widget_finder.dart';
 
 import 'platform.dart';
 import 'widget_base.dart';
+import 'platform_provider.dart';
 
 const double _kMinInteractiveDimensionCupertino = 44.0;
 
@@ -73,6 +77,9 @@ class MaterialIconButtonData extends _BaseData {
     this.constraints,
     this.splashRadius,
     this.mouseCursor,
+    this.isSelected,
+    this.selectedIcon,
+    this.style,
   });
 
   final AlignmentGeometry? alignment;
@@ -89,6 +96,9 @@ class MaterialIconButtonData extends _BaseData {
   final BoxConstraints? constraints;
   final double? splashRadius;
   final MouseCursor? mouseCursor;
+  final bool? isSelected;
+  final Widget? selectedIcon;
+  final ButtonStyle? style;
 }
 
 class PlatformIconButton extends PlatformWidgetBase<CupertinoButton, Widget> {
@@ -147,6 +157,9 @@ class PlatformIconButton extends PlatformWidgetBase<CupertinoButton, Widget> {
       constraints: data?.constraints,
       splashRadius: data?.splashRadius,
       mouseCursor: data?.mouseCursor ?? SystemMouseCursors.click,
+      isSelected: data?.isSelected,
+      selectedIcon: data?.selectedIcon,
+      style: data?.style,
     );
   }
 
@@ -157,11 +170,23 @@ class PlatformIconButton extends PlatformWidgetBase<CupertinoButton, Widget> {
     // child is required non nullable
     assert(data?.icon != null || cupertinoIcon != null || icon != null);
 
+    // If the IconButton is placed inside the AppBar, we need to have zero padding.
+    final haveZeroPadding = PlatformProvider.of(context)
+            ?.settings
+            .iosUseZeroPaddingForAppbarPlatformIcon ??
+        false;
+    final isPlacedOnPlatformAppBar =
+        ParentWidgetFinder.of<CupertinoNavigationBar>(context) != null;
+    final overriddenPadding =
+        haveZeroPadding && isPlacedOnPlatformAppBar ? EdgeInsets.zero : null;
+
+    final givenPadding = data?.padding ?? padding ?? overriddenPadding;
+
     return CupertinoButton(
       key: data?.widgetKey ?? widgetKey,
       child: data?.icon ?? cupertinoIcon ?? icon!,
       onPressed: data?.onPressed ?? onPressed ?? null,
-      padding: data?.padding ?? padding,
+      padding: givenPadding,
       color: data?.color ?? color,
       borderRadius: data?.borderRadius ??
           const BorderRadius.all(const Radius.circular(8.0)),